requirements.txt
numpy
pandas
scikit-learn
torch
rdkit-pypi
pyyaml
tqdm
matplotlib
networkx
faiss-cpu
pre-commit
